requirements.txt
amqp==2.1.4
anyjson==0.3.3
beautifulsoup4==4.6.0
billiard==3.5.0.2
bs4==0.0.1
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
configparser==3.5.0
Django==1.11.2
django-crispy-forms==1.6.1
django-formtools==2.0
django-jinja==2.3.1
django-reversion==2.0.9
django-simple-captcha==0.5.5
django-xadmin==0.5.0
future==0.16.0
html5lib==0.999999999
httplib2==0.10.3
idna==2.5
Jinja2==2.9.6
kombu==4.0.2
lxml==3.8.0
MarkupSafe==1.0
mongoelector==0.6.1
mongoengine==0.13.0
mysqlclient==1.3.10
olefile==0.44
Pillow==4.2.1
pymongo==3.4.0
pytz==2017.2
requests==2.18.1
selenium==3.4.3
six==1.10.0
urllib3==1.21.1
vine==1.1.3
webencodings==0.5.1
